requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.6
django-crispy-forms==2.3
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
